Commenting Behavior.js.  Renamed BehaviorContainer component to BehaviorInputContainer.  Updated related imports and implementations to BehaviorContainer renaming.
