+++ client/src/components/BehaviorInputContainer.jsx
